Reorganize page streaming
The "entry point" (returned by methods) now implements
IEnumerable<TResource>, and you have to ask for the paged view
explicitly.

Naming of the six interfaces involved (three for each of sync and
async) is still under discussion.
