--- src/Google.Api.Gax/PagedEnumerable.cs
@@ -1,124 +1,0 @@
-﻿/*
- * Copyright 2016 Google Inc. All Rights Reserved.
- * Use of this source code is governed by a BSD-style
- * license that can be found in the LICENSE file or at
- * https://developers.google.com/open-source/licenses/bsd
- */
-
-using Google.Protobuf;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Google.Api.Gax
-{
-    public sealed class PagedEnumerable<TRequest, TResponse, TResource> : IPagedEnumerable<TResponse, TResource>
-        where TRequest : class, IPageRequest, IMessage<TRequest>
-        where TResponse : class, IPageResponse<TResource>, IMessage<TResponse>
-    {
-        private readonly CallSettings _callSettings;
-        private readonly TRequest _request;
-        private readonly ApiCall<TRequest, TResponse> _apiCall;
-
-        public PagedEnumerable(ApiCall<TRequest, TResponse> apiCall,
-            TRequest request, CallSettings callSettings)
-        {
-            _callSettings = callSettings;
-            _request = request;
-            _apiCall = apiCall;
-        }
-
-        public IPagedEnumerator<TResponse> GetEnumerator() =>
-            new PagedEnumerator(_callSettings, _request.Clone(), _apiCall);
-
-        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
-        IEnumerator<TResponse> IEnumerable<TResponse>.GetEnumerator() => GetEnumerator();
-
-        /// <inheritdoc />
-        public IEnumerable<TResource> Flatten() => this.SelectMany(page => page);
-
-        /// <inheritdoc />
-        public IEnumerable<FixedSizePage<TResource>> WithFixedPageSize(int pageSize)
-        {
-            GaxPreconditions.CheckArgument(pageSize > 0, nameof(pageSize), "Must be greater than 0");
-            using (var enumerator = GetEnumerator())
-            {
-                bool done = false;
-                while (!done)
-                {
-                    var items = new List<TResource>(pageSize);
-                    while (items.Count < pageSize)
-                    {
-                        int requestCount = pageSize - items.Count;
-                        done = !enumerator.MoveNext(requestCount);
-                        if (done)
-                        {
-                            break;
-                        }
-                        items.AddRange(enumerator.Current);
-                        if (items.Count > pageSize)
-                        {
-                            throw new NotSupportedException("Invalid server response: " +
-                                $"requested {requestCount} items, received {enumerator.Current.Count()} items");
-                        }
-                    }
-                    if (items.Count != 0)
-                    {
-                        yield return new FixedSizePage<TResource>(items, enumerator.Current.NextPageToken);
-                    }
-                }
-            }
-        }
-
-        private class PagedEnumerator : IPagedEnumerator<TResponse>
-        {
-            private readonly CallSettings _callSettings;
-            private readonly ApiCall<TRequest, TResponse> _apiCall;
-            private readonly TRequest _request; // This is mutated during iteration
-            private bool _finished;
-
-            public PagedEnumerator(CallSettings callSettings,
-                TRequest request, ApiCall<TRequest, TResponse> apiCall)
-            {
-                _callSettings = callSettings;
-                _request = request;
-                _apiCall = apiCall;
-            }
-
-            public TResponse Current { get; private set; }
-
-            object IEnumerator.Current => Current;
-
-            public bool MoveNext()
-            {
-                if (_finished)
-                {
-                    return false;
-                }
-                Current = _apiCall.Sync(_request, _callSettings);
-                var nextPageToken = Current.NextPageToken;
-                if (nextPageToken == "")
-                {
-                    _finished = true;
-                }
-                // Prepare the next request...
-                _request.PageToken = nextPageToken;
-                return true;
-            }
-
-            public bool MoveNext(int pageSize)
-            {
-                _request.PageSize = pageSize;
-                return MoveNext();
-            }
-
-            public void Reset()
-            {
-                throw new NotImplementedException();
-            }
-
-            public void Dispose() { }
-        }
-    }
-}
